tacit-knowledge/data_acquisition/coordinator/requirements.txt
# Lambda Coordinator Dependencies
paramiko>=3.4.0
psycopg2-binary>=2.9.9
boto3>=1.34.0
requests>=2.31.0
